fix bug with deletion CardSet along with linked Cards

+++ src/main/java/com/eazylearn/controller/UserRestController.java
@@ -10,9 +10,11 @@ import com.eazylearn.security.jwt.JwtUser;
 import com.eazylearn.security.jwt.JwtUserDetailsService;
 import com.eazylearn.service.UserService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.ResponseEntity.HeadersBuilder;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,11 +25,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.ws.rs.ForbiddenException;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 import static com.eazylearn.enums.UserRole.ADMIN;
+import static com.eazylearn.security.jwt.JwtUserDetailsService.isUserHasAuthority;
 import static com.eazylearn.util.Constants.USERS_ENDPOINT_PATH;
 import static com.eazylearn.util.Convertor.uuidToString;
 import static java.util.stream.Collectors.toList;
@@ -38,6 +41,7 @@ import static org.springframework.http.ResponseEntity.ok;
 @RequestMapping(value = USERS_ENDPOINT_PATH)
 @CrossOrigin
 @RequiredArgsConstructor
+@Slf4j
 public class UserRestController { // todo: maybe endpoint reset password?
 
     private final UserService userService;
@@ -46,7 +50,7 @@ public class UserRestController { // todo: maybe endpoint reset password?
     private final JwtUserDetailsService jwtUserDetailsService;
 
     @GetMapping
-    @PreAuthorize("hasRole('ADMIN')")
+    @Secured("ROLE_ADMIN")
     public ResponseEntity<List<UserResponseDTO>> getAllUsers() {
         final List<User> allUsers = userService.findAllUsers();
 
@@ -61,16 +65,12 @@ public class UserRestController { // todo: maybe endpoint reset password?
      **/
     @GetMapping("/{id}")
     public ResponseEntity<UserResponseDTO> getUserById(@PathVariable("id") UUID userId) {
-        final JwtUser currentUser = jwtAuthenticationFacade.getJwtPrincipal();
-        if (currentUser.getId().equals(userId.toString())
-                || jwtUserDetailsService.isUserHasAuthority(currentUser, ADMIN.name())) {
-            final User user = userService.findUserById(uuidToString(userId));
+        // check if operation is allowed for current user
+        final String id = uuidToString(userId);
+        checkIfCurrentUserIdOrCurrentUserIsAdmin(id);
 
-            return ok(userMapper.toResponseDTO(user));
-        } else {
-            throw new ForbiddenException();
-        }
-
+        final User user = userService.findUserById(id);
+        return ok(userMapper.toResponseDTO(user));
     }
 
     @PostMapping
@@ -82,13 +82,31 @@ public class UserRestController { // todo: maybe endpoint reset password?
     @PatchMapping("/{id}")
     public ResponseEntity<UserResponseDTO> updateUserById(@PathVariable("id") UUID userId,
                                                           @RequestBody UserUpdateRequestDTO requestDTO) {
-        final User updatedUser = userService.updateUserById(uuidToString(userId), requestDTO);
+        // check if operation is allowed for current user
+        final String id = uuidToString(userId);
+        checkIfCurrentUserIdOrCurrentUserIsAdmin(id);
+
+        final User updatedUser = userService.updateUserById(id, requestDTO);
         return ok(userMapper.toResponseDTO(updatedUser));
     }
 
     @DeleteMapping("/{id}")
     public HeadersBuilder<?> deleteUserById(@PathVariable("id") UUID userId) {
+        // check if operation is allowed for current user
+        final String id = uuidToString(userId);
+        checkIfCurrentUserIdOrCurrentUserIsAdmin(id);
+
         userService.deleteUserById(uuidToString(userId));
         return noContent();
+    }
+
+    private void checkIfCurrentUserIdOrCurrentUserIsAdmin(String userId) {
+        final JwtUser currentUser = jwtAuthenticationFacade.getJwtPrincipal();
+        if (!Objects.equals(userId, currentUser.getId())
+                && !isUserHasAuthority(currentUser, ADMIN.toString())) {
+            log.error("User with id={} hasn't permission to perform action", currentUser.getId());
+            throw new AccessDeniedException(
+                    String.format("User with id=%s hasn't permission to perform action", currentUser.getId()));
+        }
     }
 }
